requirements.txt
Django==1.6
South==1.0.2
argparse==1.2.1
distribute==0.6.24
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap3-datetimepicker-2==2.4.2
django-datetime-widget==0.9.3
psycopg2==2.5.3
python-dateutil==2.2
pytz==2017.2
six==1.10.0
wsgiref==0.1.2
